pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "alen_enterprise"
version = "0.1.0"
description = "Information Audit Tool integrated with Gemini"
authors = ["Admin <admin@example.com>"]
readme = "README.md"
packages = [{include = "alen"}]

[tool.poetry.scripts]
alen = "alen.cli:cli"

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
httpx = "^0.27.0"
pydantic = "^2.6.3"
python-dotenv = "^1.0.1"
rich = "^13.7.1"
python-nmap = "^0.7.1"
google-generativeai = "^0.4.0"
jinja2 = "^3.1.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.2"
pytest-asyncio = "^0.23.5"
black = "^24.2.0"
isort = "^5.13.2"
flake8 = "^7.0.0"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-v --tb=short"
testpaths = [
    "tests",
]
